Renamed leds --> led_strip
