+++ Drive/src/led_strip/led_strip.h
